refactor: 이름 변경 (#215)
